backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3
python-multipart==0.0.5
